Create README.md
If you like this demo-game it would be much appreciated if you could use this link https://paypal.me/Developer001?country.x=ZA&locale.x=en_US because I want
to set up a complete version of it on Steam, but I don't have the money required for it. So any
donations would help a lot.